requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
click==8.1.7
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
frozenlist==1.4.1
fsspec==2024.2.0
greenlet==1.1.2
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.5
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.3
multidict==6.0.5
netifaces==0.11.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
pillow==10.2.0
PyGObject==3.42.1
PyJWT==2.3.0
pynvim==0.4.2
pyparsing==2.4.7
python-apt==2.4.0+ubuntu1
PyYAML==5.4.1
SecretStorage==3.3.1
six==1.16.0
sympy==1.12
systemd-python==234
torch==2.2.1
torchvision==0.17.1
triton==2.2.0
typing_extensions==4.10.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
wadllib==1.3.6
Werkzeug==3.0.1
wifite==2.6.0
yarl==1.9.4
zipp==1.0.0
